+++ packages/widgetbook/lib/src/widgetbook.dart
@@ -33,6 +33,7 @@ class Widgetbook extends StatefulWidget {
     this.themeMode,
     this.home = const DefaultHomePage(),
     this.header,
+    this.headerPadding,
     this.scrollBehavior,
     this.enableLeafComponents = true,
   });
@@ -50,6 +51,7 @@ class Widgetbook extends StatefulWidget {
     this.themeMode,
     this.home = const DefaultHomePage(),
     this.header,
+    this.headerPadding,
     this.scrollBehavior,
     this.enableLeafComponents = true,
   });
@@ -67,6 +69,7 @@ class Widgetbook extends StatefulWidget {
     this.themeMode,
     this.home = const DefaultHomePage(),
     this.header,
+    this.headerPadding,
     this.scrollBehavior,
     this.enableLeafComponents = true,
   });
@@ -120,6 +123,9 @@ class Widgetbook extends StatefulWidget {
   /// This can be used for branding or additional information.
   final Widget? header;
 
+  /// An optional padding for the header widget in the navigation panel.
+  final EdgeInsetsGeometry? headerPadding;
+
   /// The [ScrollBehavior] to be applied to the Widgetbook application itself.
   ///
   /// This allows users to override the behavior of scrolling on both desktop
@@ -150,6 +156,7 @@ class _WidgetbookState extends State<Widgetbook> {
       appBuilder: widget.appBuilder,
       home: widget.home,
       header: widget.header,
+      headerPadding: widget.headerPadding,
       addons: widget.addons,
       integrations: widget.integrations,
       enableLeafComponents: widget.enableLeafComponents,